AddressBook: documenting + refactoring. Refs #305
- Move filesystem code into new files
- Refactor layout to linearly reflect design
- Doxygen document all classes and members
- Document and comment-out unused code
- Move GetB32Address() into core/Identity.h
  - GetB32Address() is a data identity function
    that's not specific to address book
- Add many TODO's for ensuing commits
