+++ src/client/address_book.h
@@ -44,15 +44,21 @@
 #include <string>
 #include <vector>
 
+#include "address_book_storage.h"
 #include "identity.h"
 #include "core/util/filesystem.h"
+#include "destination.h"
+#include "router_context.h"
 #include "util/base64.h"
 #include "util/log.h"
 
 namespace i2p {
 namespace client {
 
+// TODO(anonimal): datatype/container refactor
+
 const char DEFAULT_SUBSCRIPTION_ADDRESS[] =
+   // TODO(anonimal): replace with monero subscription address
   "http://udhdrtrcetjm5sxzskjyr5ztpeszydbh4dpl3pl4utgqqw2v4jna.b32.i2p/hosts.txt";
 const int INITIAL_SUBSCRIPTION_UPDATE_TIMEOUT = 3;  // in minutes
 const int INITIAL_SUBSCRIPTION_RETRY_TIMEOUT = 1;  // in minutes
@@ -60,126 +66,197 @@ const int CONTINIOUS_SUBSCRIPTION_UPDATE_TIMEOUT = 720;  // in minutes (12 hours
 const int CONTINIOUS_SUBSCRIPTION_RETRY_TIMEOUT = 5;  // in minutes
 const int SUBSCRIPTION_REQUEST_TIMEOUT = 60;  // in second
 
-inline std::string GetB32Address(
-    const i2p::data::IdentHash& ident) {
-  return ident.ToBase32().append(".b32.i2p");
-}
-
-// interface for storage
-class AddressBookStorage {
- public:
-  virtual ~AddressBookStorage() {}
-
-  virtual bool GetAddress(
-      const i2p::data::IdentHash& ident,
-      i2p::data::IdentityEx& address) const = 0;
-
-  virtual void AddAddress(
-      const i2p::data::IdentityEx& address) = 0;
-
-  virtual void RemoveAddress(
-      const i2p::data::IdentHash& ident) = 0;
-
-  virtual int Load(
-      std::map<std::string,
-      i2p::data::IdentHash>& addresses) = 0;
-
-  virtual int Save(
-      const std::map<std::string, i2p::data::IdentHash>& addresses) = 0;
-};
-
-class ClientDestination;
 class AddressBookSubscription;
+/// @class AddressBook
+/// @brief Address book implementation
 class AddressBook {
  public:
-  AddressBook();
-  ~AddressBook();
+  /// @brief Initializes defaults for address book implementation
+  AddressBook()
+      : m_Storage(nullptr),
+        m_DefaultSubscription(nullptr),
+        m_SubscriptionsUpdateTimer(nullptr),
+        m_SharedLocalDestination(nullptr),
+        m_IsLoaded(false),
+        m_IsDownloading(false) {}
 
+  /// @brief Stops address book implementation
+  ~AddressBook() {
+    Stop();
+  }
+
+  /// @brief Starts address book fetching and processing of subscriptions
+  /// @param local_destination Shared pointer to destination instance used to
+  ///   generate lease set and tunnel pool for downloading new subscriptions
   void Start(
       std::shared_ptr<ClientDestination> local_destination);
 
+  /// @brief Stops and handles cleanup for address book
+  /// @details Stops timers, finishes downloading if in progress, saves
+  ///   addresses in memory to disk, cleans up memory
   void Stop();
 
+  /// @brief Gets identity hash from .b32.i2p address
+  /// @return True if identity hash was found
+  /// @param address Reference to .b32.i2p address
+  /// @param ident Reference to identity hash
+  /// @notes If identity hash is found point @param ident to found hash
   bool GetIdentHash(
       const std::string& address,
       i2p::data::IdentHash& ident);
 
-  bool GetAddress(
-      const std::string& address,
-      i2p::data::IdentityEx& identity);
-
+  /// @brief Finds address within loaded subscriptions
+  /// @returns Unique pointer to identity hash of address
+  /// @param address Reference to address
   std::unique_ptr<const i2p::data::IdentHash> FindAddress(
       const std::string& address);
 
-  std::shared_ptr<ClientDestination> GetSharedLocalDestination() const;
+  /// @brief Used for destination to fetch subscriptions
+  /// @return Shared pointer to client destination
+  std::shared_ptr<ClientDestination> GetSharedLocalDestination() const {
+    return m_SharedLocalDestination;
+  }
 
-  // for jump service
+  /// @brief Inserts address into address book from HTTP Proxy jump service
+  /// @param address Reference to human-readable address
+  /// @param base64 Reference to Base64 address
   void InsertAddress(
       const std::string& address,
       const std::string& base64);
 
-  void InsertAddress(
-      const i2p::data::IdentityEx& address);
-
+  /// @brief Loads hosts from stream into address book
+  /// @param f Reference to file stream of hosts
   void LoadHostsFromStream(
       std::istream& f);
 
+  /// @brief Sets the download state as complete and resets timer as needed
+  /// @details Resets update timer according to the state of completed download
+  /// @param success True if successful download, false if not
+  /// @notes If download was successful, reset the with regular update timeout.
+  ///   If not, reset timer with more frequent update timeout.
   void DownloadComplete(
       bool success);
 
-  // This method returns the ".b32.i2p" address
+  /// @brief Returns identity hash's .b32.i2p address
+  /// @param ident Reference to identity hash
+  /// @return Identity hash's .b32.i2p address
   std::string ToAddress(
       const i2p::data::IdentHash& ident) {
     return GetB32Address(ident);
   }
 
+  /**
+  // TODO(unassigned): currently unused
   std::string ToAddress(
       const i2p::data::IdentityEx& ident) {
     return ToAddress(ident.GetIdentHash());
   }
 
- private:
-  void StartSubscriptions();
-  void StopSubscriptions();
+  // TODO(unassigned): currently unused
+  void InsertAddress(
+      const i2p::data::IdentityEx& address);
 
-  std::unique_ptr<AddressBookStorage> CreateStorage();
-  void LoadHosts();
+  // TODO(unassigned): currently unused
+  bool GetAddress(
+      const std::string& address,
+      i2p::data::IdentityEx& identity);
+  **/
+
+ private:
+  /// @brief Starts the subscription process
+  /// @details Calls to load subscriptions, parent "poller" for fetching
+  ///   in-net subscriptions
+  void StartSubscriptions();
+
+  /// @brief Loads subscriptions
+  /// @details If not loaded, loads from file (if available)
+  ///   and instantiates an address book subscription instance
   void LoadSubscriptions();
 
+  /// @brief Implements subscription update timer
+  /// @param ecode Reference to boost error code
+  /// @details If subscriptions are available, load hosts into memory,
+  ///   otherwise retry request for subscriptions until they are downloaded
   void HandleSubscriptionsUpdateTimer(
       const boost::system::error_code& ecode);
 
+  /// @brief Loads hosts.txt
+  /// @details If not on filesystem, downloads from subscription address(es)
+  void LoadHosts();
+
+  /// @brief Creates new instance of storage for address book
+  /// @return Unique pointer to interface for address book storage
+  std::unique_ptr<AddressBookStorage> CreateStorage();
+
+  /// @brief Stop subscription process
+  /// @details Cancels subscription updates
+  void StopSubscriptions();
+
  private:
+  /// @var m_AddressBookMutex
+  /// @brief Mutex for address book implementation when loading hosts from file
   std::mutex m_AddressBookMutex;
-  std::map<std::string, i2p::data::IdentHash>  m_Addresses;
+
+  /// @var m_Addresses
+  /// @brief Map of human readable addresses to identity hashes
+  std::map<std::string, i2p::data::IdentHash> m_Addresses;
+
+  /// @var m_Storage
+  /// @brief Unique pointer to address book storage implementation
   std::unique_ptr<AddressBookStorage> m_Storage;
-  volatile bool m_IsLoaded, m_IsDownloading;
+
+  /// @var m_Subscriptions
+  /// @brief Vector of unique pointers to respective subscription implementation
   std::vector<std::unique_ptr<AddressBookSubscription>> m_Subscriptions;
 
-  // in case if we don't know any addresses yet
+  /// @var m_DefaultSubscription
+  /// @brief Unique pointer to address book subscription implementation
+  /// @notes Used if we don't have any subscription addresses yet
   std::unique_ptr<AddressBookSubscription> m_DefaultSubscription;
 
+  /// @var m_SubscriptionsUpdateTimer
+  /// @brief Unique pointer to Boost.Asio deadline_timer
+  /// @details Handles all timer-related needs for subscription fetching
   std::unique_ptr<boost::asio::deadline_timer> m_SubscriptionsUpdateTimer;
+
+  /// @var m_SharedLocalDestination
+  /// @brief Shared pointer to client destination instance
+  /// @notes Needed for fetching subscriptions in-net
   std::shared_ptr<ClientDestination> m_SharedLocalDestination;
+
+  // TODO(anonimal): the use of volatile appears to act as both an
+  // MSVC extension and substitute for std::atomic.
+  // Refactor with std::atomic.
+  volatile bool m_IsLoaded, m_IsDownloading;
 };
 
+/// @class AddressBookSubscription
+/// @brief Handles fetching of hosts.txt subscriptions
 class AddressBookSubscription {
  public:
+  /// @brief Initializes defaults for address book subscription implementation
   AddressBookSubscription(
       AddressBook& book,
-      const std::string& link);
+      const std::string& link)
+      : m_Book(book),
+        m_Link(link) {}
 
+  /// @brief Instantiates thread that fetches in-net subscriptions
   void CheckSubscription();
 
  private:
+  /// @warning Must be run in separate thread
   void Request();
 
  private:
+  /// @var m_Book
+  /// @brief Reference to address book implementation
   AddressBook& m_Book;
+  // Used for HTTP request  // TODO(anonimal): remove when refactored with cpp-netlib
   std::string m_Link, m_Etag, m_LastModified;
 };
 
 }  // namespace client
 }  // namespace i2p
 
-#endif  // SRC_CLIENT_ADDRESSBOOK_H_
+#endif  // SRC_CLIENT_ADDRESS_BOOK_H_
